requirements.txt
autopep8==1.5.6
pycodestyle==2.7.0
pypiwin32==223
python-dotenv==0.17.0
pywin32==300
selenium==3.141.0
toml==0.10.2
urllib3==1.26.4
win10toast==0.9
